+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * elevation: fixes on lineargradient (https://github.com/Raruto/leaflet-elevation/issues/306#issuecomment-4420914053)
 * listmarker: works on multiple maps on one page now
+* custom icons in media library for gpx, kml, geojson
 
 Please note the [install instructions](https://leafext.de/en/doku/about/versions/).